dist-foreign-ui/src/main/java/com/copyright/rup/dist/foreign/ui/usage/impl/sal/ViewSalFundPoolMediator.java
package com.copyright.rup.dist.foreign.ui.usage.impl.sal;

import com.copyright.rup.dist.foreign.ui.main.security.ForeignSecurityUtils;
import com.copyright.rup.vaadin.widget.api.IMediator;

import com.vaadin.ui.Button;

/**
 * Mediator for the view SAL fund pool widget.
 * <p/>
 * Copyright (C) 2020 copyright.com
 * <p/>
 * Date: 09/25/2020
 *
 * @author Aliaksandr Liakh
 */
class ViewSalFundPoolMediator implements IMediator {

    private Button deleteButton;

    @Override
    public void applyPermissions() {
        deleteButton.setVisible(ForeignSecurityUtils.hasDeleteFundPoolPermission());
    }

    public void setDeleteButton(Button deleteButton) {
        this.deleteButton = deleteButton;
    }
}
